requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
pandas
numpy
scikit-learn
joblib
psycopg2-binary
gunicorn
python-dotenv
